requirements.txt
pyaudio
pygame
pydub
obs-websocket-py
requests